Initial script for 3D visualization of proteins.
This is an html page that incorporates iCn3D Structure viewer (https://www.ncbi.nlm.nih.gov/Structure/icn3d/icn3d.html#bHighlight). I'm using this because it seems to have a nice UI and functions already built in. The current version of the script just loads HA (1RUZ) from PDB and zooms in on a set of random residues.