add in code for preferences histogram
Still more work to be done here in order to properly load data from JSONs, but at least now we have spec'd out a selectable histogram with site details on demand. Next thing to do is to implement with grouping to allow summing of preferences for multiple amino acids tolerated at one site.