Convert the requested site to an integer for frequency lookup
The site sent to the frequencies panel from the line plot is sent as a string,
but it needs to be converted to an integer to lookup the corresponding
frequencies at the site. With this commit, one can click on a circle in the line
plot panel and the frequencies panel will update for the selected site.