Rename global variables and disable zoom box
The zoom box that enabled zoom interactives with the focus plot was being
plotted on top of the frequencies panel because both line and frequencies plots
used the same global variable name `svg` and a race condition led to the zoom
rect being appended in the frequencies svg. This commit renames the global
variables to address this conflict, but it also disables the zoom box for now to
ensure that tooltips and interactions with individual circles on the line plot
work.

Closes #13.