Fix frequencies legend loading and position
Moved frequencies legend just outside the data display region on the righthand
side to prevent data from obscuring legend values. Also, fixed logic for
displaying the legend so each change of site creates a new legend instance and
color scale. This change ensures that the legend updates when the site changes
and also that the legend does not show a different color for all possible amino
acid mutations.