Resize panels to fit on a single screen with minimal HTML/CSS
This commit merges recent changes to the protein display and also adds a
new_index.html file that lays out the figure panels with Bootstrap CSS and a
minimal HTML file. Several experimental features were removed in the process
including the modal button and color picker. Also commented out in this commit
the key features for interacting with the protein structure including saving
image, loading a PDB file, etc. These will be fixed in a subsequent commit.