Emphasize points in line plot with frequency data
Attempts to address problem where points with frequency data (shown in black)
were occluded by points without data by reducing the opacity and size of points
without data and also the line connecting points.

The ideal way to address this problem with SVG is to plot the points with data
after the points without data, but this would require more refactoring of our
code than we have time for now.