Select the site with the highest antigenic escape value by default
This fixes an issue where the default frequencies shown were not related to what
was shown in the line plot or protein structure and also provides the user with
a meaningful default display when the page loads.

This commit adds some logic to the protein structure javascript to select the
site with the highest immune escape value. Since the "click" events in the line
plot can't happen until the protein structure is loaded, the simulated click
event needs to happen at the end of the protein structure load code. This logic
is supported by adding an id to each circle in the line plot based on the
circle's site value. Then we can easily select the site with the highest immune
escape and dispatch a click event there to load the corresponding frequencies
and show the position on the protein structure.