Setup axes as in line plot and name SVG groups for stacked bars
This commit fixes the interactive selection of points in the line plot and
corresponding display of stacked bars for the selected sites. The main
improvement here is the naming of the SVG groups in the barchart to distinguish
them from other "g" tags that were getting selected prior to the d3 join and
causing the join to fail with obscure errors. The moral of the story: always
name your groups.