added selection by brushing and clicking
Now you can select a point by brushing or by clicking.
You can deselect a point by clicking.

The current bug is that if you click and then brush, the
clicked point will become unselected.
I think this is because the brush function sets the selected class
to only the points in the brush box.
This is not true of brushing and then clicking.
In this case, the brushed points remain selected.