Move brushing function outside of updateChart
Fixes an issue where brushing in the focus panel was broken on the first load of
the app unless the updateChart function was called twice. This turned out to be
a race-condition where the circlePoint was not properly populated when the
brushing function was defined. Calling updateChart the second time redefined the
brushing function with a fully populated circlePoint structure, allowing
brushing to work.

The solution here was to move the brushing function out of the updateChart
function and select circles from the focus panel with a simple d3 select when
the function was called. This change removes the brushing function's dependency
on an internal variable of updateChart and fixes the bug.

Closes #47.