Update context plot on dataset change
Changes the d3 logic for creating the context plot to use the `data` function
and also a `join` call. This approach allows the context line to be redrawn when
the dataset changes.

This commit also changes the logic for how the line plot is drawn in the focus
panel, using the same data and join approach. While this change mostly fixes the
line plot in the focus panel, there is an edge case where a dataset change with
a zoomed-in view of the data redraws the entire gene's data without resizing to
the appropriate context of the current zoom. For some reason, the circles redraw
for new data without any issues, but the line does not.

Closes #48.