Display mutation metrics in logoplots based on dropdown menu
Adds a dropdown menu for selecting any available mutation metric from the same
inputu data used to plot the site-level metrics and updates the logic of the
logoplots to draw the appropriate data for the selected condition and mutation
metric.

As part of these changes, this commit also removes a redundant promise from the
main app script and adds new variables to the line plot chart including
`mutData` and `condition_mut_data` attributes that parallel the site-specific
data stored in the existing `data` and `condition_data` attributes.

This commit also introduces an updated IAV CSV file with a fake mutation metric,
"mut_fakesel", whose values are those of the real "mut_diffsel" column after
random shuffling. This additional metric allows us to test the effects of
switching between metrics from the dropdown menu.