removed debounce functionality
Previously, the function to select brushed points was 'debounced.'
This means that it was only called a maximum of 1 time per
pre-defined period of time. This was because when the brush box
is adjusted, the selection function is called many times.

However, the new brush strategy removes these calls. Because the
brush box cannot be adjusted, the selection function is only called
once per "swipe." Therefore, we do not need to debounce the
selection function.