Scale widths of logoplot letters as more letters are added
This is a partial solution to handling selection of "too many" sites in the line
plot for display in the logoplots. The solution is inspired by standard plotting
behavior for barcharts where the bandwidth of each bar gets smaller as more
sites are added. One difference here is that the scaling only occurs when the
bandwidth becomes smaller than the letter width. If we scale letters _up_ to
match the bandwidth, they become illegible.

There are some remaining issues called here by TODOs where widths are
hardcoded. These issues should be resolve as part of a broader cleanup effort
with respect to how letters are generated and how their widths/heights are
calculated. Specifically, we really should only need to generate letters and
calculate their dimensions once and then we can reuse those values latter.

Closes #60.