Dynamically load data defined by user config in front matter
Load data and PDB paths from variables in the front matter of individual
markdown pages, create a single page per dataset, and configure Jekyll to render
these pages along with a list of available datasets both in the index and at the
bottom of each dataset page.

One side effect of moving dataset pages into a subdirectory is that URLs for
styles and fonts need to be absolute instead of relative.