Render user-defined Markdown to describe the data being viewed
Provides a very rough proof-of-concept for allowing users to provide their data
in separate pieces. In this case, users can provide a URL to a Markdown
document (e.g., "h3n2_markdown.md") that the app will fetch and render at the
bottom of the page.

The app will also track the Markdown URL in the query string and check for a
Markdown URL in the query string on load. This is a related approach to letting
users provide data while also maintaining the state of the app in the URL. This
similar approach could be augmented with drag-and-drop functionality and/or the
ability to specify URLs for the CSV and PDB files.