Add text fields for data, PDB, and Markdown URLs
In addition to these new text fields, this commit changes the behavior of the
existing Markdown URL field such that any change to the field triggers a fetch
of the corresponding URL instead of relying on a submit button. If the fetch
fails (as in a 404 error), the text field is highlighted with a Bootstrap form
CSS class to provide some feedback that the provided value is invalid.

If a Markdown URL is provided in the query string when the page is loaded, that
value is also used to populate the corresponding text field.