Fix bug where brushing broke after switching datasets
After switching between datasets (e.g., HIV to IAV), the brush selection of
sites didn't work as expected. Sometimes no sites would be selected and
sometimes completely different sites than the ones being brushed would be
selected. The problem was caused by the fact that d3 joins use the index of the
given arrays as the id of each record, by default [1]. This means that entry 10
of the HIV data mapped to entry 10 of the IAV data when these latter data were
loaded, even though the "site id" of those records were completely different.

This commit fixes the problem by explicitly reassigning the CSS id for each
circle in the site plot (e.g., "#site_123") during the enter and update steps of
the join. A better solution might be to define a more dataset-specific key for
the joins, so d3 could distinguish between a record from HIV or IAV data.

[1] https://github.com/d3/d3-selection/blob/master/README.md#selection_data