Update circle color on d3 update
Since joins already maintain the "selected" state of each circle, we don't need
to call "selectSite" again. However, one side effect of re-selecting sites was
to update their color. This commit applies a function to the d3 update step to
assign the proper color to each selected circle.

Note that this fill update has to happen before the transition animations or the
selection of the site with the max metric value doesn't work as expected and the
max site remains gray.