Specify default data to load when none are provided by the user
This loads the app with default data instead of showing empty panels. We may
want to change this approach in the future, since it forces data to be loaded
for all three panels. Users can currently workaround this by providing an empty
query string for any component they don't want to display. For example, if a
user wanted to omit the protein view, they could include "&pdb-url=" in the URL
and the app would treat this not as missing data but as a request for an empty
URL.