Prevent logos from extending to the left of the x-axis origin
When too many sites are selected in the site plot, the logoplot columns would
get drawn to the left of the x-axis origin and on top of the y-axis line. This
occurred because the repositioning of letters based on bandwidth and letter
width was only necessary when the letter width was less than the bandwidth.

In addition to fixing this primary issue, this commit also replaces the
hardcoded letter width of 60 with a dynamically calculated width that should be
the same for all letters of a given font size (based on the "advance width" of
the letter). This width is only calculated once prior to plotting and then
reused for all subsequent transforms.

Closes #105.