Limit the number of labels plotted per site in the logoplot
Fixes an issue where the selection of too many sites in the site plot resulted
in drawing tick labels for all of those sites in the logoplot. These labels were
unreadable because they were too densely packed. This commit introduces two
variables to control the maximum number of sites for which all labels should be
drawn and then a number of sites to skip between each label.

This solution may address most use cases for our current data, but a more robust
solution would take into account the actual width of the label text trying to be
displayed and then downsample the labels intelligently as the density
increases. We may need to look at the implementation of d3's axis ticks [1] for
inspiration.

Closes #89.

[1] https://observablehq.com/@d3/axis-ticks