Track the URL for data displayed in the site plot
Fixes a bug associated with running a d3 join based on the default key of the
array index where switching from a dataset with more sites to a dataset with
fewer caused the site plot to show circles in the wrong x axis position.

This commit adds a "data_url" entry to each data record and tracks the URL of
the currently loaded dataset. This information is used to generate an id for
each record going into the d3 join where the id is the data URL plus the site
id. When a new dataset is loaded, the data URL is updated and the d3 join
performs enter/exit operations instead of the update operation it was previously
using. One side effect of this approach is that selected sites are also cleared
when a new dataset is loaded.

This commit includes a related change that clears the brush in the context view
when switching datasets. This clearing prevents previously zoomed in views from
appearing to be broken when the focus view updates.

Closes #118.