Fix display of tooltips in site and logo plots
Fixes a bug where hovering over points in the site plot inconsistently showed
the tooltip and also somehow made clicking on individual points difficult. This
commit follows advice from Mila Frerichs [1] by appending the tooltip div to the
body of the page instead of the SVG element and using the d3 event page
coordinates instead of the mouse position coordinates to place the
tooltips.

Uses the same approach to fix the positioning of the logo plot tooltip such that
it is closer to the letter being hovered over.

Closes #110.

[1] https://mappingwithd3.com/tutorials/basics/tooltip/