Track state of dropdown fields in the URL
Adds two new functions to set the state of the dropdowns from the given URL and
also to update the URL when the state of the dropdowns changes.

Also removes a call to "updateChart" from the initial site plot creation and
uses a call to "dropdownChange" instead to get the same effect based on the
state of the dropdowns. This may not be an appropriate solution and needs to be
tested more.