Do not allow users to specify invalid field values in the app state
Prevents bugs when switching between datasets when the URL requests a specific
value for a dropdown (e.g., condition or site metric) that doesn't exist for the
new dataset. The app now issues a warning on the console. This should be a
warning through a unified feedback mechanism in the app, though.

TODO: Make a unified feedback mechanism for the app.