Track selected sites in URL and form field
Adds ability to specify selected sites in the URL and in a form field near the
"clear selections" button. Selecting sites by clicking or brushing, etc. updates
the form field accordingly and changing the contents of the field manually
updates the sites that are selected. If sites are specified in the URL at load,
those sites are selected instead of the site with the maximum value. As selected
sites change, the URL is updated, too.

Although this implementation works, it feels gross. There is a lot of reaching
between main and line_plot_zoom js files using the DOM and I had to add a
promise to get things to work. There must be a better way to do this kind of
thing and the answer might be something like React et al.