Only select max site when no selected sites are passed by URL
Allows the user to pass an empty list of selected sites in the URL such that
opening that URL will cause no sites to be selected. This fixes an edge case bug
where an empty string in the selected sites URL parameter was ignored and the
site with the maximum value was selected instead.