Pin versions for opentype.js and marked
Fixes an issue that caused our site to not return promises as expected after
loading fonts. This issue was caused by updates to opentype.js and our loading
of the latest version of that library. This commit fixes the issue by pinning
opentype.js to an earlier version prior to the broken API. This commit also
pins the version of the marked library we use for markdown parsing, to
future-proof our tool to similar breaking changes in the future.