Add basic ability to download SVGs from D3 panels
Adds download buttons to the line and logo plot panels and links these to a
downloadSVG function that serializes the SVG XML of each panel and sends it to
the user as a downloaded file. This commit does not add the ability to download
the PDB panel. The PDB panel uses Canvas instead of SVG, so we need to use NGL's
built-in mechanisms or another approach to support downloads.

Related to #147.