Track protein color options in the URL
Adds logic to check for color values in protein color fields and display those
colors on the protein structure when the app loads. Also adds colors to the URL
as they are selected.

This commit also fixes a bug when the PDB URL is provided and the data URL is
not provided. In this setting, the app checks for whether the chart variable is
undefined or not before trying to color chains on the protein. The chart
variable is always defined, though, even if data are not loaded for it, and as a
result the protein view tries to color chains it doesn't have information
about. This commit changes the check for whether the `chart.data` attribute is
undefined instead.